Fixed CUDA warnings

+++ encog-cmd/cuda_test.c
@@ -120,8 +120,8 @@ void TestCUDA()
         printf("  GPU Clock Speed:                               %.2f GHz\n", deviceProp.clockRate * 1e-6f);
   
     
-        printf("  Total amount of constant memory:               %u bytes\n", deviceProp.totalConstMem); 
-        printf("  Total amount of shared memory per block:       %u bytes\n", deviceProp.sharedMemPerBlock);
+        printf("  Total amount of constant memory:               %u bytes\n", (unsigned int)deviceProp.totalConstMem); 
+        printf("  Total amount of shared memory per block:       %u bytes\n", (unsigned int)deviceProp.sharedMemPerBlock);
         printf("  Total number of registers available per block: %d\n", deviceProp.regsPerBlock);
         printf("  Warp size:                                     %d\n", deviceProp.warpSize);
         printf("  Maximum number of threads per block:           %d\n", deviceProp.maxThreadsPerBlock);
@@ -133,8 +133,8 @@ void TestCUDA()
                deviceProp.maxGridSize[0],
                deviceProp.maxGridSize[1],
                deviceProp.maxGridSize[2]);
-        printf("  Maximum memory pitch:                          %u bytes\n", deviceProp.memPitch);
-		printf("  Texture alignment:                             %u bytes\n", deviceProp.textureAlignment);
+        printf("  Maximum memory pitch:                          %u bytes\n", (unsigned int)deviceProp.memPitch);
+		printf("  Texture alignment:                             %u bytes\n", (unsigned int)deviceProp.textureAlignment);
 }
 	printf("Performing CUDA test.\n");
 	TestVectorAdd();
